Consolidated swap components out into new components.
