+++ src/components/CardHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CardHeader'
