+++ src/components/OptionTextInfo/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OptionTextInfo'
